examples/requirements.txt
# Plot
matplotlib
numpy
scikit-learn

# Explore
dash
plotly
umap-learn